requirements.txt
asgiref==3.8.1
Django==5.1.1
djangorestframework==3.15.2
mysqlclient==2.2.4
psycopg2-binary==2.9.9
python-decouple==3.8
sqlparse==0.5.1
